+++ src/AbstractDefinition.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace HbLib\Container;
+
+use Closure;
+use ReflectionFunction;
+
+abstract class AbstractDefinition
+{
+}